Move tests to *.spec.js (#261)
* Move tests to *.spec.js

* Update .codeclimate.yml
