+++ .codeclimate.yml
@@ -12,4 +12,5 @@ ratings:
   - "**.jsx"
 exclude_paths:
 - "test"
-- "**/__tests__"
+- "**/*.spec.js"
+- "**/*.spec.jsx"
