+++ src/CenturyView/Decade.spec.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import Day from '../Day';
+import Decade from './Decade';
 
 /* eslint-disable comma-dangle, jsx-a11y/mouse-events-have-key-events */
 
 const tileProps = {
   activeStartDate: new Date(2018, 0, 1),
   classes: ['react-calendar__tile'],
-  currentMonthIndex: 0,
-  date: new Date(2018, 0, 1),
+  date: new Date(2011, 0, 1),
+  point: 2011,
 };
 
-describe('Day', () => {
+describe('Decade', () => {
   it('applies given classNames properly', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         classes={['react-calendar__tile', 'react-calendar__tile--flag']}
         tileClassName={() => 'testFunctionClassName'}
@@ -26,93 +26,67 @@ describe('Day', () => {
 
     expect(wrapperClassName.includes('react-calendar__tile')).toBe(true);
     expect(wrapperClassName.includes('react-calendar__tile--flag')).toBe(true);
-    expect(wrapperClassName.includes('react-calendar__month-view__days__day')).toBe(true);
+    expect(wrapperClassName.includes('react-calendar__century-view__decades__decade')).toBe(true);
     expect(wrapperClassName.includes('testFunctionClassName')).toBe(true);
   });
 
-  it('applies additional classNames for weekends', () => {
+  it('renders component without abbreviation', () => {
     const component = mount(
-      <Day
-        {...tileProps}
-        date={new Date(2018, 0, 6)} // Saturday
-      />
-    );
-
-    const wrapperClassName = component.find('.react-calendar__tile').prop('className');
-
-    expect(wrapperClassName.includes('react-calendar__month-view__days__day--weekend')).toBe(true);
-  });
-
-  it('applies additional classNames for neighboring months', () => {
-    const component = mount(
-      <Day
-        {...tileProps}
-        date={new Date(2018, 1, 2)}
-      />
-    );
-
-    const wrapperClassName = component.find('.react-calendar__tile').prop('className');
-
-    expect(wrapperClassName.includes('react-calendar__month-view__days__day--neighboringMonth')).toBe(true);
-  });
-
-  it('renders component with proper abbreviation', () => {
-    const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         date={new Date(2018, 0, 1)}
+        decade={2011}
       />
     );
 
     const abbr = component.find('abbr');
 
-    expect(abbr).toHaveLength(1);
-    expect(abbr.prop('aria-label')).toBe('January 1, 2018');
-    expect(component.text()).toBe('1');
+    expect(abbr).toHaveLength(0);
+    expect(component.text()).toBe('2011 – 2020');
   });
 
-  it('is disabled when date is before beginning of minDate\'s day', () => {
+  it('is disabled when date is before beginning of minDate\'s decade', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
-        date={new Date(2018, 0, 1)}
-        minDate={new Date(2018, 0, 2)}
+        date={new Date(2010, 0, 1)}
+        minDate={new Date(2020, 0, 1)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeTruthy();
   });
 
-  it('is not disabled when date is after beginning of minDate\'s day', () => {
+  it('is not disabled when date is after beginning of minDate\'s decade', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
-        date={new Date(2018, 0, 1)}
-        minDate={new Date(2018, 0, 1)}
+        date={new Date(2010, 0, 1)}
+        minDate={new Date(2010, 0, 1)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeFalsy();
   });
 
-  it('is disabled when date is after end of maxDate\'s day', () => {
+  it('is disabled when date is after end of maxDate\'s decade', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
-        date={new Date(2018, 0, 2)}
-        maxDate={new Date(2018, 0, 1)}
+        date={new Date(2020, 0, 1)}
+        maxDate={new Date(2010, 0, 1)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeTruthy();
   });
 
-  it('is not disabled when date is before end of maxDate\'s day', () => {
+  it('is not disabled when date is before end of maxDate\'s decade', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
-        date={new Date(2018, 0, 1)}
-        maxDate={new Date(2018, 0, 1)}
+        date={new Date(2010, 0, 1)}
+        maxDate={new Date(2010, 0, 1)}
       />
     );
 
@@ -120,11 +94,11 @@ describe('Day', () => {
   });
 
   it('calls onClick callback when clicked and sends proper date as an argument', () => {
-    const date = new Date(2018, 0, 1);
+    const date = new Date(2010, 0, 1);
     const onClick = jest.fn();
 
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         date={date}
         onClick={onClick}
@@ -138,11 +112,11 @@ describe('Day', () => {
   });
 
   it('calls onMouseOver callback when hovered and sends proper date as an argument', () => {
-    const date = new Date(2018, 0, 1);
+    const date = new Date(2010, 0, 1);
     const onMouseOver = jest.fn();
 
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         date={date}
         onMouseOver={onMouseOver}
@@ -156,11 +130,11 @@ describe('Day', () => {
   });
 
   it('calls onMouseOver callback when focused and sends proper date as an argument', () => {
-    const date = new Date(2018, 0, 1);
+    const date = new Date(2010, 0, 1);
     const onMouseOver = jest.fn();
 
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         date={date}
         onMouseOver={onMouseOver}
@@ -175,7 +149,7 @@ describe('Day', () => {
 
   it('renders tileContent properly', () => {
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         tileContent={<div className="testContent" />}
       />
@@ -192,7 +166,7 @@ describe('Day', () => {
     tileContent.mockReturnValue(<div className="testContent" />);
 
     const component = mount(
-      <Day
+      <Decade
         {...tileProps}
         date={date}
         tileContent={tileContent}
@@ -202,7 +176,7 @@ describe('Day', () => {
     const testContent = component.find('.testContent');
 
     expect(tileContent).toHaveBeenCalled();
-    expect(tileContent).toHaveBeenCalledWith({ date, view: 'month' });
+    expect(tileContent).toHaveBeenCalledWith({ date, view: 'century' });
     expect(testContent).toHaveLength(1);
   });
 });
