+++ src/MonthView/Day.spec.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import Decade from '../Decade';
+import Day from './Day';
 
 /* eslint-disable comma-dangle, jsx-a11y/mouse-events-have-key-events */
 
 const tileProps = {
   activeStartDate: new Date(2018, 0, 1),
   classes: ['react-calendar__tile'],
-  date: new Date(2011, 0, 1),
-  point: 2011,
+  currentMonthIndex: 0,
+  date: new Date(2018, 0, 1),
 };
 
-describe('Decade', () => {
+describe('Day', () => {
   it('applies given classNames properly', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         classes={['react-calendar__tile', 'react-calendar__tile--flag']}
         tileClassName={() => 'testFunctionClassName'}
@@ -26,67 +26,93 @@ describe('Decade', () => {
 
     expect(wrapperClassName.includes('react-calendar__tile')).toBe(true);
     expect(wrapperClassName.includes('react-calendar__tile--flag')).toBe(true);
-    expect(wrapperClassName.includes('react-calendar__century-view__decades__decade')).toBe(true);
+    expect(wrapperClassName.includes('react-calendar__month-view__days__day')).toBe(true);
     expect(wrapperClassName.includes('testFunctionClassName')).toBe(true);
   });
 
-  it('renders component without abbreviation', () => {
+  it('applies additional classNames for weekends', () => {
     const component = mount(
-      <Decade
+      <Day
+        {...tileProps}
+        date={new Date(2018, 0, 6)} // Saturday
+      />
+    );
+
+    const wrapperClassName = component.find('.react-calendar__tile').prop('className');
+
+    expect(wrapperClassName.includes('react-calendar__month-view__days__day--weekend')).toBe(true);
+  });
+
+  it('applies additional classNames for neighboring months', () => {
+    const component = mount(
+      <Day
+        {...tileProps}
+        date={new Date(2018, 1, 2)}
+      />
+    );
+
+    const wrapperClassName = component.find('.react-calendar__tile').prop('className');
+
+    expect(wrapperClassName.includes('react-calendar__month-view__days__day--neighboringMonth')).toBe(true);
+  });
+
+  it('renders component with proper abbreviation', () => {
+    const component = mount(
+      <Day
         {...tileProps}
         date={new Date(2018, 0, 1)}
-        decade={2011}
       />
     );
 
     const abbr = component.find('abbr');
 
-    expect(abbr).toHaveLength(0);
-    expect(component.text()).toBe('2011 – 2020');
+    expect(abbr).toHaveLength(1);
+    expect(abbr.prop('aria-label')).toBe('January 1, 2018');
+    expect(component.text()).toBe('1');
   });
 
-  it('is disabled when date is before beginning of minDate\'s decade', () => {
+  it('is disabled when date is before beginning of minDate\'s day', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
-        date={new Date(2010, 0, 1)}
-        minDate={new Date(2020, 0, 1)}
+        date={new Date(2018, 0, 1)}
+        minDate={new Date(2018, 0, 2)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeTruthy();
   });
 
-  it('is not disabled when date is after beginning of minDate\'s decade', () => {
+  it('is not disabled when date is after beginning of minDate\'s day', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
-        date={new Date(2010, 0, 1)}
-        minDate={new Date(2010, 0, 1)}
+        date={new Date(2018, 0, 1)}
+        minDate={new Date(2018, 0, 1)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeFalsy();
   });
 
-  it('is disabled when date is after end of maxDate\'s decade', () => {
+  it('is disabled when date is after end of maxDate\'s day', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
-        date={new Date(2020, 0, 1)}
-        maxDate={new Date(2010, 0, 1)}
+        date={new Date(2018, 0, 2)}
+        maxDate={new Date(2018, 0, 1)}
       />
     );
 
     expect(component.find('.react-calendar__tile').prop('disabled')).toBeTruthy();
   });
 
-  it('is not disabled when date is before end of maxDate\'s decade', () => {
+  it('is not disabled when date is before end of maxDate\'s day', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
-        date={new Date(2010, 0, 1)}
-        maxDate={new Date(2010, 0, 1)}
+        date={new Date(2018, 0, 1)}
+        maxDate={new Date(2018, 0, 1)}
       />
     );
 
@@ -94,11 +120,11 @@ describe('Decade', () => {
   });
 
   it('calls onClick callback when clicked and sends proper date as an argument', () => {
-    const date = new Date(2010, 0, 1);
+    const date = new Date(2018, 0, 1);
     const onClick = jest.fn();
 
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         date={date}
         onClick={onClick}
@@ -112,11 +138,11 @@ describe('Decade', () => {
   });
 
   it('calls onMouseOver callback when hovered and sends proper date as an argument', () => {
-    const date = new Date(2010, 0, 1);
+    const date = new Date(2018, 0, 1);
     const onMouseOver = jest.fn();
 
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         date={date}
         onMouseOver={onMouseOver}
@@ -130,11 +156,11 @@ describe('Decade', () => {
   });
 
   it('calls onMouseOver callback when focused and sends proper date as an argument', () => {
-    const date = new Date(2010, 0, 1);
+    const date = new Date(2018, 0, 1);
     const onMouseOver = jest.fn();
 
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         date={date}
         onMouseOver={onMouseOver}
@@ -149,7 +175,7 @@ describe('Decade', () => {
 
   it('renders tileContent properly', () => {
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         tileContent={<div className="testContent" />}
       />
@@ -166,7 +192,7 @@ describe('Decade', () => {
     tileContent.mockReturnValue(<div className="testContent" />);
 
     const component = mount(
-      <Decade
+      <Day
         {...tileProps}
         date={date}
         tileContent={tileContent}
@@ -176,7 +202,7 @@ describe('Decade', () => {
     const testContent = component.find('.testContent');
 
     expect(tileContent).toHaveBeenCalled();
-    expect(tileContent).toHaveBeenCalledWith({ date, view: 'century' });
+    expect(tileContent).toHaveBeenCalledWith({ date, view: 'month' });
     expect(testContent).toHaveLength(1);
   });
 });
