+++ src/YearView.spec.jsx
@@ -1,32 +1,33 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import { getBeginOfDecade, getEndOfDecade } from '../shared/dates';
-
-import CenturyView from '../CenturyView';
+import YearView from './YearView';
 
 /* eslint-disable comma-dangle, react/prop-types */
 
-describe('CenturyView', () => {
+const { format } = new Intl.DateTimeFormat('en-US', { month: 'long', year: 'numeric' });
+
+describe('YearView', () => {
   it('renders proper view when given activeStartDate', () => {
-    const activeStartDate = new Date(2001, 0, 1);
+    const activeStartDate = new Date(2017, 0, 1);
     const component = mount(
-      <CenturyView
+      <YearView
         activeStartDate={activeStartDate}
         showNeighboringMonth={false}
       />
     );
 
     const firstDayTile = component.find('.react-calendar__tile').first();
+    const firstDayTileTimeAbbr = firstDayTile.find('abbr').prop('aria-label');
 
-    expect(firstDayTile.text()).toBe(`${getBeginOfDecade(activeStartDate).getFullYear()} – ${getEndOfDecade(activeStartDate).getFullYear()}`);
+    expect(firstDayTileTimeAbbr).toBe(format(activeStartDate));
   });
 
   it('applies tileClassName to its tiles when given a string', () => {
-    const activeStartDate = new Date(2001, 0, 1);
+    const activeStartDate = new Date(2017, 0, 1);
     const tileClassName = 'testClassName';
     const component = mount(
-      <CenturyView
+      <YearView
         activeStartDate={activeStartDate}
         showNeighboringMonth={false}
         tileClassName={tileClassName}
@@ -40,7 +41,7 @@ describe('CenturyView', () => {
   });
 
   it('applies tileClassName to its tiles conditionally when given a function that returns a string', () => {
-    const activeStartDate = new Date(2001, 0, 1);
+    const activeStartDate = new Date(2017, 0, 1);
     const tileClassNameFn = ({ date }) => {
       if (date.getTime() === activeStartDate.getTime()) {
         return 'firstDayOfTheMonth';
@@ -49,7 +50,7 @@ describe('CenturyView', () => {
       return null;
     };
     const component = mount(
-      <CenturyView
+      <YearView
         activeStartDate={activeStartDate}
         showNeighboringMonth={false}
         tileClassName={tileClassNameFn}
@@ -68,12 +69,12 @@ describe('CenturyView', () => {
   });
 
   it('renders tileContent in its tiles when given a node', () => {
-    const activeStartDate = new Date(2001, 0, 1);
+    const activeStartDate = new Date(2017, 0, 1);
     const tileContent = (
       <div className="testContent" />
     );
     const component = mount(
-      <CenturyView
+      <YearView
         activeStartDate={activeStartDate}
         showNeighboringMonth={false}
         tileContent={tileContent}
@@ -89,7 +90,7 @@ describe('CenturyView', () => {
   });
 
   it('renders tileContent in its tiles conditionally when given a function that returns a node', () => {
-    const activeStartDate = new Date(2001, 0, 1);
+    const activeStartDate = new Date(2017, 0, 1);
     const tileContentFn = ({ date }) => {
       if (date.getTime() === activeStartDate.getTime()) {
         return (
@@ -100,7 +101,7 @@ describe('CenturyView', () => {
       return null;
     };
     const component = mount(
-      <CenturyView
+      <YearView
         activeStartDate={activeStartDate}
         showNeighboringMonth={false}
         tileContent={tileContentFn}
@@ -116,5 +117,20 @@ describe('CenturyView', () => {
 
     expect(firstDayTileContent).toHaveLength(1);
     expect(secondDayTileContent).toHaveLength(0);
+  });
+
+  it('displays year view with custom month formatting', () => {
+    const activeStartDate = new Date(2017, 0, 1);
+
+    const component = mount(
+      <YearView
+        activeStartDate={activeStartDate}
+        formatMonth={() => 'Month'}
+      />
+    );
+
+    const month = component.find('.react-calendar__year-view__months__month').first();
+
+    expect(month.text()).toBe('Month');
   });
 });
